Use akitaPreUpdateEntity to format contract versions
