+++ libs/contract/src/lib/+state/contract.firestore.ts
@@ -18,11 +18,6 @@ export enum ContractStatus {
   aborted = 'abordted',
 }
 
-/** An interface for a single document to display versions subcollection count. */
-export interface VersionMeta extends ContractVersionDocumentWithDates{
-  count: number;
-}
-
 export interface ContractTitleDetail {
   /**
    * @dev titleId is replacing movieId
