+++ libs/contract/src/lib/version/+state/contract-version.model.ts
@@ -2,6 +2,7 @@ import { createTerms } from "@blockframes/utils/common-interfaces/terms";
 import { createPrice } from "@blockframes/utils/common-interfaces/price";
 import { createContract, Contract } from "../../+state/contract.model";
 import { ContractVersionDocumentWithDates, ContractStatus } from "../../+state/contract.firestore";
+import { firestore } from "firebase/app";
 
 export type ContractVersion = ContractVersionDocumentWithDates;
 
@@ -12,6 +13,11 @@ export type ContractVersion = ContractVersionDocumentWithDates;
 export interface ContractWithLastVersion {
   doc: Contract,
   last: ContractVersion,
+}
+
+/** An interface for a single document to display versions subcollection count. */
+export interface VersionMeta extends ContractVersionDocumentWithDates{
+  count: number;
 }
 
 export function createContractVersion(params: Partial<ContractVersion> = {}): ContractVersion {
@@ -31,4 +37,20 @@ export function initContractWithVersion(): ContractWithLastVersion {
     doc: createContract(),
     last: createContractVersion(),
   }
+}
+
+/**
+ *
+ * @param contractVersion
+ */
+export function formatContractVersion(contractVersion: any): ContractVersion {
+  // Dates from firebase are Timestamps, we convert it to Dates.
+  if (contractVersion.scope && contractVersion.scope.start instanceof firestore.Timestamp) {
+    contractVersion.scope.start = contractVersion.scope.start.toDate();
+  }
+
+  if (contractVersion.scope.end instanceof firestore.Timestamp) {
+    contractVersion.scope.end = contractVersion.scope.end.toDate();
+  }
+  return contractVersion as ContractVersion;
 }
