+++ libs/contract/src/lib/version/+state/contract-version.store.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ContractVersion } from './contract-version.model';
 import { EntityState, ActiveState, EntityStore, StoreConfig } from '@datorama/akita';
-import { firestore } from 'firebase-admin';
+import { formatContractVersion } from './contract-version.model';
 
 export interface ContractVersionState extends EntityState<ContractVersion, string>, ActiveState<string> {}
 
@@ -14,14 +14,7 @@ export class ContractVersionStore extends EntityStore<ContractVersionState> {
   }
 
   akitaPreAddEntity(contractVersion: ContractVersion) {
-    if (contractVersion.scope && contractVersion.scope.start instanceof firestore.Timestamp) {
-      contractVersion.scope.start = contractVersion.scope.start.toDate();
-    }
-
-    if (contractVersion.scope.end instanceof firestore.Timestamp) {
-      contractVersion.scope.end = contractVersion.scope.end.toDate();
-    }
-    return contractVersion as ContractVersion;
+    return formatContractVersion(contractVersion);
   }
 
 }
